+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.1] - 2023-11-29
+
+- Fixed "sTestFile" pattern matching skipping every other entry.
+
 ## [0.6.0] - 2023-11-15
 
 - Added ASI Loader installer